--- a/.env.example
+++ b/.env.example
@@ -1,0 +1,52 @@
+# DankerChat Environment Configuration
+# Copy this file to .env and update with your actual values
+
+# Flask Configuration
+FLASK_ENV=development
+FLASK_DEBUG=true
+FLASK_SECRET_KEY=your-secret-key-here
+
+# Database Configuration
+DATABASE_URL=sqlite:///dankerchat.db
+# For production PostgreSQL:
+# DATABASE_URL=postgresql://user:password@localhost:5432/dankerchat
+
+# Redis Configuration (for sessions and WebSocket)
+REDIS_URL=redis://localhost:6379/0
+
+# JWT Configuration
+JWT_SECRET_KEY=your-jwt-secret-key-here
+JWT_ACCESS_TOKEN_EXPIRES=3600
+
+# Application Settings
+APP_NAME=DankerChat
+APP_VERSION=0.1.0
+DEBUG=true
+
+# Server Configuration
+HOST=0.0.0.0
+PORT=5000
+
+# WebSocket Configuration
+SOCKETIO_ASYNC_MODE=threading
+
+# Logging
+LOG_LEVEL=DEBUG
+LOG_FILE=logs/dankerchat.log
+
+# UV-specific Environment Variables
+# These are automatically handled by UV but can be customized
+
+# UV_PYTHON_PREFERENCE=only-managed  # Use only UV-managed Python versions
+# UV_SYSTEM_PYTHON=false             # Don't use system Python
+# UV_COMPILE_BYTECODE=true           # Compile .pyc files for faster startup
+# UV_PROJECT_ENVIRONMENT=.venv       # Virtual environment location (default)
+
+# Development Tools Configuration
+COVERAGE_RCFILE=pyproject.toml
+MYPY_CONFIG_FILE=pyproject.toml
+PYTEST_ADDOPTS=--strict-markers --strict-config -v
+
+# Testing Database (separate from development)
+TEST_DATABASE_URL=sqlite:///:memory:
+TEST_REDIS_URL=redis://localhost:6379/1