--- a/.gitignore
+++ b/.gitignore
@@ -1,0 +1,91 @@
+# DankerChat .gitignore
+# Generated for UV-based Python development
+
+# Python
+__pycache__/
+*.py[cod]
+*$py.class
+*.so
+*.egg
+*.egg-info/
+dist/
+build/
+eggs/
+parts/
+var/
+sdist/
+develop-eggs/
+.installed.cfg
+
+# UV Virtual Environment
+.venv/
+# Note: uv.lock should be committed for reproducible builds
+
+# IDEs and Editors
+.vscode/
+.idea/
+*.swp
+*.swo
+*~
+
+# Testing and Coverage
+.pytest_cache/
+.coverage
+htmlcov/
+.tox/
+.nox/
+coverage.xml
+*.cover
+.hypothesis/
+
+# Environment Variables
+.env
+.env.local
+.env.development
+.env.test
+.env.production
+
+# Logs
+*.log
+logs/
+
+# Database
+*.db
+*.sqlite
+*.sqlite3
+
+# OS Files
+.DS_Store
+Thumbs.db
+
+# Temporary Files
+*.tmp
+*.temp
+/tmp/
+
+# Build artifacts
+*.tar.gz
+*.zip
+
+# Documentation builds
+docs/_build/
+site/
+
+# Jupyter Notebook
+.ipynb_checkpoints
+
+# Flask instance folder
+instance/
+
+# Redis dump file
+dump.rdb
+
+# Backup files created during migration
+backup-requirements.txt
+environment-snapshot.txt
+scripts-backup/
+
+# Development artifacts
+.mypy_cache/
+.ruff_cache/
+.black_cache/